+++ src/challenges/js/30DaysJs-29/src/exercise.js
@@ -10,33 +10,3 @@ export class BasicProduct extends Product {
     return this.description;
   }
 }
-
-export class WarrantyDecorator extends Product {
-  constructor(product) {
-    super(product);
-    this.product = product;
-  }
-
-  getPrice() {
-    return this.product.getPrice() + 20;
-  }
-
-  getDescription() {
-    return `${this.product.getDescription()} con garantía`;
-  }
-}
-
-export class ShippingInsuranceDecorator extends Product {
-  constructor(product) {
-    super(product);
-    this.product = product;
-  }
-
-  getPrice() {
-    return this.product.getPrice() + 20;
-  }
-
-  getDescription() {
-    return `${this.product.getDescription()} con seguro de envío`;
-  }
-}
